Cleaned up /ui/shell/navBar/ and /ui/layout/ file structure.
