+++ packages/app/client/src/ui/shell/main.tsx
@@ -38,11 +38,11 @@ import * as React from 'react';
 import { Colors, Fonts, Splitter } from '@bfemulator/ui-react';
 import ExplorerBar from './explorer';
 import MDI from './mdi';
-import NavBar from './navBar';
+import { NavBar } from './navBar';
 import { TabManager, DialogHost } from '../dialogs';
 import * as Constants from '../../constants';
-import StatusBar from './statusBar';
-import StoreVisualizer from '../debug/storeVisualizer';
+import { StatusBar } from './statusBar';
+import { StoreVisualizer } from '../debug/storeVisualizer';
 import store from '../../data/store';
 import { KeyCodes } from '@uifabric/utilities/lib';
 import * as PresentationActions from '../../data/action/presentationActions';
