+++ packages/app/client/src/ui/shell/statusBar.tsx
@@ -31,4 +31,19 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-export * from './genericDocument';
+import { css } from 'glamor';
+import * as React from 'react';
+import { Colors } from '@bfemulator/ui-react';
+
+const CSS = css({
+  backgroundColor: Colors.STATUS_BAR_BACKGROUND_DARK,
+  color: Colors.STATUS_BAR_FOREGROUND_DARK,
+  lineHeight: '22px',
+  minHeight: '22px',
+  width: '100%',
+  alignSelf: 'flex-end',
+  flex: 1
+});
+
+export const StatusBar = props =>
+  <div { ...CSS }></div>
